requirements.txt
Flask
Werkzeug
python-docx
pdfminer.six
scikit-learn
pandas
google-generativeai
python-dotenv
